Updated frontend controllers, fixed Backend building entity.
Api models:
- Added new constructors for ease of use in the frontend.
- Changed convertModel() to be able to serialize nulls.

Frontend controllers:
- Updated method calls to use new building entity.

Backend:
- Changed any reference from 'Integer id' to 'Long id'
- Added proper endpoint for insert/new_building

+++ server/src/main/java/nl/tudelft/oopp/server/services/BuildingService.java
@@ -28,7 +28,7 @@ public class BuildingService {
      * @param id we search for
      * @return the building with that exact id if it exists or null if not
      */
-    public Optional<Building> getBuilding(Integer id) {
+    public Optional<Building> getBuilding(Long id) {
         return buildingRepository.findById(id);
     }
 
@@ -47,7 +47,7 @@ public class BuildingService {
      * @param id       new one to be updated to
      * @param building to be updated
      */
-    public void updateBuilding(Integer id, Building building) {
+    public void updateBuilding(Long id, Building building) {
         buildingRepository.save(building);
     }
 
@@ -56,7 +56,7 @@ public class BuildingService {
      *
      * @param id to be deleted from the list of buildings
      */
-    public void delete(Integer id) {
+    public void delete(Long id) {
         buildingRepository.deleteById(id);
     }
 }
